Use dynamic URL for tdb if no tdb server provided
Previously the local dashboard builds were hard-compiled
to use localhost:6363. This creates the problem that if
the user chooses a different port, the dashboard will
stop working as it still tries to connect to localhost:6363
instead. Therefore I use the window.url.origin to determine
where the dashboard is served if the ENV variable is not
provided.